reorganized server files under src directory and reorganized index.js and app.js
